% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{.cut}
\alias{.cut}
\title{.Cut}
\usage{
.cut(ordered_result = TRUE, ...)
}
\arguments{
\item{ordered_result}{\code{bool} If \code{TRUE}, the output will be an ordered factor. Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to \link{cut}.}
}
\description{
\link{cut} with sensible defaults for binning continuous variables : includes the lowest
value observed and closed on the left. By default the result will be ordered.
}
