% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{.summarize}
\alias{.summarize}
\title{.Summarize}
\usage{
.summarize(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \link{summarize}.}
}
\description{
Tidyverse \link{summarize} that drops groups by default.
}
