% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_date.R
\name{excel_to_date}
\alias{excel_to_date}
\title{Convert Excel Garbage into an Actual Date}
\usage{
excel_to_date(date)
}
\arguments{
\item{date}{\code{str/num} Date code to be converted}
}
\description{
Convert excel date codes into human (and R) readable dates.
}
\examples{
date <- 8573
excel_to_date(date)

date <- '8573'
excel_to_date(date)

}
