% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french_to_numeric.R
\name{french_to_numeric}
\alias{french_to_numeric}
\title{Convert french-style number string to numeric}
\usage{
french_to_numeric(str)
}
\arguments{
\item{str}{\code{chr} String to convert.}
}
\description{
Takes a french style string to a numberic, for example \code{"1 001,5"} becomes \code{1001.5}.
}
\examples{
french_to_numeric("1 001,5")

}
