% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_week_start.R
\name{get_week_start}
\alias{get_week_start}
\title{Start of Week}
\usage{
get_week_start(week, year)
}
\arguments{
\item{week}{\code{int/str} ISO week number, zero padding is accepted but not required.}

\item{year}{\code{int/str} ISO year. Must be formatted as YYYY.}
}
\description{
Gets the first monday of a given ISO week-year.
}
