% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{\code{value / list} Value(s) to look for in \code{y}.}

\item{y}{\code{list} List of values.}
}
\value{
\verb{\%notin\%} operator
}
\description{
Convenience operator inverting the functionality of \code{\%in\%}; ie, x \%notin\% y will
return a bool (or vector of bools) indicating whether x (or its members) are in y.
}
\examples{
\dontrun{
  test_strings <- c('hello', 'world')
  c('foobar', 'hello') \%notin\% test_strings
}

}
