% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_outbreaks.R
\name{identify_outbreaks}
\alias{identify_outbreaks}
\title{Identify Outbreaks}
\usage{
identify_outbreaks(
  df,
  min_weekly = 20,
  min_size = 100,
  min_duration = 6,
  max_gap = 16,
  tail_threshold = 5,
  tail_tolerance = 4
)
}
\arguments{
\item{df}{\code{dataframe} Historical data to be analyzed. note, this function has \strong{strong}
expectations of column names.}

\item{min_weekly}{\code{int} Number of minimum cases expected per week during the 'main phase' of the
outbreak. Note that gaps of up to \code{max_gap} in length are tolerated.}

\item{min_size}{\code{int} Minimum number of cases required during the 'main phase' of the epidemic.}

\item{min_duration}{\code{int} Minimum duration (in weeks) required during the 'main phase' of the
epidemic.}

\item{max_gap}{\code{int} Maximum lull (in weeks) during which an epidemic can dip below the expected
20 case per week threshold. Gaps longer than this period would indicate that there were two
seperate epidemics.}

\item{tail_threshold}{\code{int} Threshold of weekly cases needed for inclusion within the 'tails' of
of the epidemic.}

\item{tail_tolerance}{\code{int} Minimum number of weeks with an incidence below \code{tail_threshold}
required for the tails to end.}
}
\value{
df
}
\description{
Returns a new version of df with two new columns: "outbreak" and "outbreak_id". The
the first is a bool indicating if an outbreak is present at that time point and the second is
an alphanumeric id with the format "{health-zone-name}_{number}" where the number indicates if it
is the first, second, etc outbreak identified (within the overall data) for that health zone.
Arguments are available to play with the identification criteria being used. Please refer to
the methods section of the final report for more information on how the identification algorithm
works.
}
