% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr.R
\name{iqr}
\alias{iqr}
\title{IQR}
\usage{
iqr(v, digits = 1, percent = FALSE, brackets = FALSE)
}
\arguments{
\item{v}{\code{vector} vector to analyze.}

\item{digits}{\code{int} number of digits after the decimal to show in the iqr values.}

\item{percent}{\code{bool} whether to display values as percents}

\item{brackets}{\code{bool} if TRUE result will have square brackets (ie: [lower - upper] instead of
lower - upper). default is FALSE.}
}
\value{
\code{str}
}
\description{
convenience function to extract the iqr of a vector as a formatted string (with or
without square brackets).
}
