% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_to_date.R
\name{iso_to_date}
\alias{iso_to_date}
\title{ISO Week-Year to Date}
\usage{
iso_to_date(
  year = format(Sys.Date(), "\%Y\%"),
  week = format(Sys.Date(), "\%V"),
  day = 1
)
}
\arguments{
\item{year}{\code{int} Year (with \%Y\% format). Default is the current year.}

\item{week}{\code{int} ISO week number (\%V). Must be between 1 and 53. A warning will be provided if
\code{week} is 53 but \code{year} is not a leap year. Default is the current week.}

\item{day}{\code{int} ISO week day. Must be between 1 and 7 where 1 is Monday and 7 is Sunday. Default
is 1 (Monday).}
}
\description{
Provides the date associated with a given ISO year - week number - day of week.
}
