% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_index.R
\name{run_index}
\alias{run_index}
\title{Run Index}
\usage{
run_index(x, FUN, n = 3, direction = c("forward", "backward"))
}
\arguments{
\item{x}{\code{vector} vector to parse.}

\item{FUN}{\code{function} function indicating the condition to be satisfied.}

\item{n}{\code{integer} minimum run length (number of TRUE values in a row). default is 3.}

\item{direction}{\code{string} whether to parse x forward or backwards (resulting in the first or last
time a run was observed respectively). default is forward.}
}
\value{
\code{int}
}
\description{
provides the index of the first time a logical condition is satisfied a minimum of
n times in a row.
}
