% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_format.R
\name{si_format}
\alias{si_format}
\title{SI prefix notation for large numers}
\usage{
si_format(x)
}
\arguments{
\item{x}{\code{dbl / int} Numer to be abreviated.}
}
\value{
\code{chr}
}
\description{
Applies SI prefix notation to a numer, \code{x}. This is intended for epidemic estimates
and works for k and M only.
}
\examples{
x <- 123456
si_format(x)

y <- 12345678
si_format(y)

}
