% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_display.R
\name{str_to_display}
\alias{str_to_display}
\title{Create Display Version of String}
\usage{
str_to_display(text)
}
\arguments{
\item{text}{\code{str} Text to be converted.}
}
\description{
Creates a display version of a string by replacing underscores with spaces and
applying title case.
}
\examples{
text <- 'bas_uele'
str_to_display(text)

}
