% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up_down_format.R
\name{up_down_format}
\alias{up_down_format}
\title{Up / down formatting for a number}
\usage{
up_down_format(
  x,
  flip_colors = FALSE,
  percent = TRUE,
  decimals = 0,
  parentheses = TRUE
)
}
\arguments{
\item{x}{\code{dbl} Number to be formatted.}

\item{flip_colors}{\code{bool} If TRUE, the default color scheme will be reversed such that negative
values are red and positive values are blue. Default is FALSE.}

\item{percent}{\code{bool} If TRUE, x will be interpreted as a proportion (ie 0.56) and displayed as a
percent (ie. 56\%). Default is TRUE.}

\item{decimals}{\code{int} Number of decimals to include in \code{x}. Default is 0.}

\item{parentheses}{\code{bool} If TRUE, the text will be wrapped in parentheses. Default is TRUE.}
}
\value{
\code{chr}
}
\description{
Creates an html string providing french format and color to x as well as adding an
up / down arrow. By default, x will be red when positive, grey when 0, and blue when negative,
but this can be inverted. Wrapping the text in parentheses is optional.
}
\examples{
x <- 5
y <- -10.76

up_down_format(x)
up_down_format(x,
               flip_colors = TRUE,
               parentheses = FALSE)
up_down_format(y,
               percent = FALSE,
               decimals = 1)

}
\seealso{
\code{\link[=french_format]{french_format()}}
}
