% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_names.R
\name{validate_names}
\alias{validate_names}
\title{Validate Names}
\usage{
validate_names(df_new, df_ref, fix_zone = NULL, fix_reg = NULL)
}
\arguments{
\item{df_new}{\code{dataframe} Dataframe with names to be validated / fixed.}

\item{df_ref}{\code{dataframe} Reference dataframe containing the accepted names you want to match to.}

\item{fix_zone}{\code{list} Named list of correct zone names (where zone is the smaller geographic
level). If  provided, this list will be used to update \code{df_new} and must have the format where
the list  names (keys) match what is in df_new and the list values are the updated (correct)
names. The default is \code{NULL}.}

\item{fix_reg}{\code{list} Named list of correct region names (where region is the larger geographic
level). Equivalent of \code{fix_zone} but at the region level. Default is \code{NULL}.}
}
\description{
Utility function to assist name validation / fixing. The function will update names
with a user provided list at both the zone (smaller) and region (larger) levels if provided. It
will also print a list of all remaining unmatched / invalid names. \strong{WARNING} : this function is
\emph{opinionated} and assumes that both \code{df_new} and \code{df_ref} have columns called "zone" and "reg";
if those columns are not present the function will fail.
}
